Create TratokBatch.sol
During testing it became apparent that as the network grows and bookings increase exponentially, there is a bottleneck when there are multiple transfers. This has been evident during testing of the 4th generation ecosystem.
Therefore this smart contract is being deployed to test on the Sepolia testnet to benchmark for performance improvements.